support clone graph without config param  (#1709)

+++ hugegraph-core/src/main/java/com/baidu/hugegraph/util/ConfigUtil.java
@@ -19,17 +19,19 @@
 
 package com.baidu.hugegraph.util;
 
+import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.file.Paths;
 import java.util.List;
 import java.util.Map;
 
 import org.apache.commons.configuration.ConfigurationException;
+import org.apache.commons.configuration.PropertiesConfiguration;
 import org.apache.commons.configuration.tree.ConfigurationNode;
-import org.apache.commons.io.Charsets;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.tinkerpop.gremlin.util.config.YamlConfiguration;
@@ -44,14 +46,16 @@ public final class ConfigUtil {
     private static final Logger LOG = Log.logger(ConfigUtil.class);
 
     private static final String NODE_GRAPHS = "graphs";
-    private static final String SUFFIX = ".properties";
+    private static final String CONF_SUFFIX = ".properties";
+    private static final String CHARSET = "UTF-8";
 
     public static void checkGremlinConfig(String conf) {
         YamlConfiguration yamlConfig = new YamlConfiguration();
         try {
             yamlConfig.load(conf);
         } catch (ConfigurationException e) {
-            throw new HugeException("Failed to load yaml config file ", conf);
+            throw new HugeException("Failed to load yaml config file '%s'",
+                                    conf);
         }
         List<ConfigurationNode> nodes = yamlConfig.getRootNode()
                                                   .getChildren(NODE_GRAPHS);
@@ -67,20 +71,20 @@ public final class ConfigUtil {
     }
 
     public static Map<String, String> scanGraphsDir(String graphsDirPath) {
-        LOG.info("Scaning graphs configuration directory {}", graphsDirPath);
+        LOG.info("Scaning option 'graphs' directory '{}'", graphsDirPath);
         File graphsDir = new File(graphsDirPath);
         E.checkArgument(graphsDir.exists() && graphsDir.isDirectory(),
-                        "Please ensure the graphs config directory '%s' " +
-                        "exist and indeed a directory", graphsDir);
+                        "Please ensure the path '%s' of option 'graphs' " +
+                        "exist and it's a directory", graphsDir);
         File[] confFiles = graphsDir.listFiles((dir, name) -> {
-            return name.endsWith(SUFFIX);
+            return name.endsWith(CONF_SUFFIX);
         });
         E.checkNotNull(confFiles, "graph configuration files");
         Map<String, String> graphConfs = InsertionOrderUtil.newMap();
         for (File confFile : confFiles) {
-            // NOTE: file name as graph name
+            // NOTE: use file name as graph name
             String name = StringUtils.substringBefore(confFile.getName(),
-                                                      ConfigUtil.SUFFIX);
+                                                      ConfigUtil.CONF_SUFFIX);
             HugeFactory.checkGraphName(name, confFile.getPath());
             graphConfs.put(name, confFile.getPath());
         }
@@ -90,24 +94,42 @@ public final class ConfigUtil {
     public static void writeToFile(String dir, String graphName,
                                    HugeConfig config) {
         E.checkArgument(FileUtils.getFile(dir).exists(),
-                        "The graphs conf directory must exist");
-        String fileName = Paths.get(dir, graphName + SUFFIX).toString();
+                        "The directory '%s' must exist", dir);
+        String fileName = Paths.get(dir, graphName + CONF_SUFFIX).toString();
         try (OutputStream os = new FileOutputStream(fileName)) {
-            config.save(os, Charsets.UTF_8.name());
+            config.save(os, CHARSET);
             config.setFileName(fileName);
-            LOG.info("Write HugeConfig to file {}", fileName);
+            LOG.info("Write HugeConfig to file: '{}'", fileName);
         } catch (IOException | ConfigurationException e) {
-            throw new HugeException("Failed to write HugeConfig to file {}",
+            throw new HugeException("Failed to write HugeConfig to file '%s'",
                                     e, fileName);
         }
     }
 
     public static void deleteFile(File file) {
+        if (!file.exists()) {
+            return;
+        }
         try {
             FileUtils.forceDelete(file);
         } catch (IOException e) {
-            throw new HugeException("Failed to delete HugeConfig file {}",
+            throw new HugeException("Failed to delete HugeConfig file '%s'",
                                     e, file);
         }
+    }
+
+    public static PropertiesConfiguration buildConfig(String configText) {
+        E.checkArgument(StringUtils.isNotEmpty(configText),
+                        "The config text can't be null or empty");
+        PropertiesConfiguration propConfig = new PropertiesConfiguration();
+        try {
+            InputStream in = new ByteArrayInputStream(configText.getBytes(
+                                                      CHARSET));
+            propConfig.setDelimiterParsingDisabled(true);
+            propConfig.load(in);
+        } catch (Exception e) {
+            throw new IllegalStateException("Failed to read config options", e);
+        }
+        return propConfig;
     }
 }
